Flush stdout
Without this, messages "building XYZ" can appear at the end of the CI
log, after all error messages (if there are any). This can be quite
confusing. The reason is the stdout is fully buffered if not connected
to a terminal and error messages go from child processes, therefore
using another buffer.